+++ ext/TensorBoardLoggerExt/TensorBoardLoggerExt.jl
@@ -4,7 +4,7 @@ using RxInfer
 using ReactiveMP: event_name, getdata
 using ExponentialFamily:
     UnivariateNormalDistributionsFamily, GammaDistributionsFamily
-using Distributions: UnivariateDistribution, Beta, Bernoulli, Binomial, InverseGamma, Poisson, Geometric, NegativeBinomial, Exponential, VonMises, Weibull, LogNormal, Erlang, shape, rate, scale, params, succprob, ntrials
+using Distributions: UnivariateDistribution, Beta, Bernoulli, Binomial, InverseGamma, Poisson, Geometric, NegativeBinomial, Exponential, VonMises, Weibull, LogNormal, Erlang, Laplace, shape, rate, scale, params, succprob, ntrials, location
 using Dates: now, format
 using Random: MersenneTwister
 using Statistics: mean, var
@@ -206,6 +206,17 @@ function _log_posterior_scalars!(
     step = (ctx.posterior_step[name] = get(ctx.posterior_step, name, 0) + 1)
     TensorBoardLogger.log_value(
         ctx.logger, "posteriors/$(name)/shape", shape(dist); step = step
+    )
+    TensorBoardLogger.log_value(
+        ctx.logger, "posteriors/$(name)/scale", scale(dist); step = step
+    )
+end
+function _log_posterior_scalars!(
+    ctx::LogContext, dist::Laplace, name::Symbol
+)
+    step = (ctx.posterior_step[name] = get(ctx.posterior_step, name, 0) + 1)
+    TensorBoardLogger.log_value(
+        ctx.logger, "posteriors/$(name)/location", location(dist); step = step
     )
     TensorBoardLogger.log_value(
         ctx.logger, "posteriors/$(name)/scale", scale(dist); step = step